requirements.txt
asgiref==3.8.1
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.6
django-cors-headers==4.4.0
google-api-core==2.19.1
google-auth==2.31.0
google-cloud-core==2.4.1
google-cloud-speech==2.26.0
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.62.2
idna==3.7
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydub==0.25.1
requests==2.32.3
rsa==4.9
sqlparse==0.5.0
urllib3==2.2.2
